[Fix] #11 - 바로구매 오류 수정
